Migrate listpack.c unit tests to new framework
