+++ src/listpack.h
@@ -97,8 +97,4 @@ lpNextRandom(unsigned char *lp, unsigned char *p, unsigned int *index, unsigned 
 int lpSafeToAdd(unsigned char *lp, size_t add);
 void lpRepr(unsigned char *lp);
 
-#ifdef SERVER_TEST
-int listpackTest(int argc, char *argv[], int flags);
-#endif
-
 #endif
